Add Location query relocation
