+++ packages/invoices/src/types.ts
@@ -8,7 +8,7 @@ import {
   Name,
   InvoiceLineNode,
 } from '@openmsupply-client/common';
-import { Location } from '@openmsupply-client/system';
+import { LocationRowFragment } from '@openmsupply-client/system';
 
 /**
  * Invoice, InvoiceRow and InvoiceLine extend the GQL types, mostly. GQL types for related entities
@@ -27,7 +27,7 @@ export interface InvoiceLine
   stockLineId: string;
   unitName?: string;
   invoiceId: string;
-  location?: Location | null;
+  location?: LocationRowFragment | null;
 }
 
 export interface InvoiceRow
