+++ packages/system/src/Location/index.ts
@@ -1,3 +1,3 @@
 export * from './Components';
-export * from './types';
 export * from './Service';
+export * from './api';
