--- packages/system/src/Location/types.ts
@@ -1,3 +1,0 @@
-import { LocationNode } from '@openmsupply-client/common';
-
-export type Location = Omit<LocationNode, '__typename' | 'stock'>;
